progress

+++ src/ui/components.h
@@ -21,7 +21,7 @@ void UI_Battery(uint8_t Level);
 void UI_RSSIBar(int16_t rssi, uint8_t line);
 void UI_F(uint32_t f, uint8_t line);
 void UI_FSmall(uint32_t f);
-void UI_DrawScrollBar(const uint8_t size, const uint8_t currentIndex,
+void UI_DrawScrollBar(const uint16_t size, const uint16_t currentIndex,
                       const uint8_t linesCount);
 void UI_ShowMenuItem(uint8_t line, const char *name, bool isCurrent);
 void UI_ShowMenu(const MenuItem *items, uint8_t size, uint8_t currentIndex);
